Refactor the git status parser to not use a state machine.
